requirements.txt
python-telegram-bot[job-queue]==21.6
feedparser>=6.0.8
requests==2.32.3
python-dotenv==1.0.1
anthropic>=0.40.0
google-genai>=1.0.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
playwright>=1.40.0
playwright-stealth>=1.0.6
cloudscraper>=1.2.71
httpx>=0.24.0
